# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Elevator Configuration Engine",
    "description": "Rule-Based Elevator Configuration and Pricing Engine (CPQ). Configures elevators, validates engineering constraints, resolves component dependencies, and calculates dynamic pricing.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/configurations/search": {
      "get": {
        "tags": [
          "Configuration",
          "Configuration"
        ],
        "summary": "Search configurations",
        "description": "Case-insensitive search by project name, display ID, or configuration ID.",
        "operationId": "search_configurations_api_v1_configurations_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Search term",
              "default": "",
              "title": "Q"
            },
            "description": "Search term"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APISuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/configurations": {
      "post": {
        "tags": [
          "Configuration",
          "Configuration"
        ],
        "summary": "Create a new configuration",
        "operationId": "create_configuration_api_v1_configurations_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConfigurationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APISuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Configuration",
          "Configuration"
        ],
        "summary": "List configurations",
        "operationId": "list_configurations_api_v1_configurations_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APISuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/configurations/{configuration_id}": {
      "get": {
        "tags": [
          "Configuration",
          "Configuration"
        ],
        "summary": "Get configuration",
        "operationId": "get_configuration_api_v1_configurations__configuration_id__get",
        "parameters": [
          {
            "name": "configuration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Configuration Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APISuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Configuration",
          "Configuration"
        ],
        "summary": "Update configuration",
        "operationId": "update_configuration_api_v1_configurations__configuration_id__put",
        "parameters": [
          {
            "name": "configuration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Configuration Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConfigurationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APISuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Configuration",
          "Configuration"
        ],
        "summary": "Delete configuration",
        "operationId": "delete_configuration_api_v1_configurations__configuration_id__delete",
        "parameters": [
          {
            "name": "configuration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Configuration Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/configurations/{configuration_id}/validate": {
      "post": {
        "tags": [
          "Configuration",
          "Configuration"
        ],
        "summary": "Run validation pipeline",
        "operationId": "validate_configuration_api_v1_configurations__configuration_id__validate_post",
        "parameters": [
          {
            "name": "configuration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Configuration Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APISuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/configurations/{configuration_id}/pricing": {
      "get": {
        "tags": [
          "Configuration",
          "Configuration"
        ],
        "summary": "Get pricing summary",
        "operationId": "get_configuration_pricing_api_v1_configurations__configuration_id__pricing_get",
        "parameters": [
          {
            "name": "configuration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Configuration Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APISuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/configurations/{configuration_id}/validation": {
      "get": {
        "tags": [
          "Configuration",
          "Configuration"
        ],
        "summary": "Get validation results",
        "operationId": "get_configuration_validation_api_v1_configurations__configuration_id__validation_get",
        "parameters": [
          {
            "name": "configuration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Configuration Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APISuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/configurations/{configuration_id}/bom": {
      "get": {
        "tags": [
          "Configuration",
          "Configuration"
        ],
        "summary": "Get bill of materials",
        "operationId": "get_configuration_bom_api_v1_configurations__configuration_id__bom_get",
        "parameters": [
          {
            "name": "configuration_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Configuration Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APISuccessEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APISuccessEnvelope": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "title": "Data"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "data",
          "timestamp"
        ],
        "title": "APISuccessEnvelope",
        "description": "Standardized API Success Response Envelope."
      },
      "CreateConfigurationRequest": {
        "properties": {
          "project_name": {
            "type": "string",
            "minLength": 1,
            "title": "Project Name",
            "description": "Customer-given project name (e.g. 'My Apartment')"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name",
            "description": "Project-in Charge / Customer Name"
          },
          "selected_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Category",
            "description": "The chosen ElevatorCategory ID"
          },
          "customer_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Reference",
            "description": "Deprecated: Optional CRM/ERP reference ID"
          }
        },
        "type": "object",
        "required": [
          "project_name"
        ],
        "title": "CreateConfigurationRequest",
        "description": "Payload for starting a new configuration session."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpdateConfigurationRequest": {
        "properties": {
          "project_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Name",
            "description": "Update project name"
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name",
            "description": "Update customer name"
          },
          "selected_feature_options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Feature Options",
            "description": "IDs of selected FeatureOptions"
          },
          "selected_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Category",
            "description": "The chosen ElevatorCategory ID"
          }
        },
        "type": "object",
        "title": "UpdateConfigurationRequest",
        "description": "Payload for updating an existing configuration."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}